fix: use integer pence for monetary values instead of float64
Floating-point arithmetic produces rounding errors that are unacceptable
for financial calculations. Store all amounts as int64 minor units (pence)
in the domain and DB layers, converting at the HTTP boundary only.

Made-with: Cursor

+++ internal/accounts/accounts.go
@@ -16,7 +16,7 @@ type BankAccount struct {
 	SortCode      string
 	Name          string
 	AccountType   string
-	Balance       float64
+	Balance       int64 // minor units (pence)
 	Currency      string
 	CreatedAt     time.Time
 	UpdatedAt     time.Time
@@ -33,7 +33,7 @@ type Transaction struct {
 	ID            string
 	AccountNumber string
 	UserID        string
-	Amount        float64
+	Amount        int64 // minor units (pence)
 	Currency      string
 	Type          string
 	Reference     *string
@@ -44,7 +44,7 @@ type CreateTransactionParams struct {
 	ID            string
 	AccountNumber string
 	UserID        string
-	Amount        float64
+	Amount        int64 // minor units (pence)
 	Currency      string
 	Type          string
 	Reference     string
